+++ packages/charts/react-charting/src/components/DeclarativeChart/DeclarativeChart.tsx
@@ -4,7 +4,7 @@ import { useTheme } from '@fluentui/react';
 import { IRefObject } from '@fluentui/react/lib/Utilities';
 import { DonutChart } from '../DonutChart/index';
 import { VerticalStackedBarChart } from '../VerticalStackedBarChart/index';
-import { PlotData, PlotlySchema } from './PlotlySchema';
+import type { PlotData, PlotlySchema } from '@fluentui/chart-utilities';
 import {
   isArrayOrTypedArray,
   isDateArray,
@@ -100,7 +100,23 @@ export const DeclarativeChart: React.FunctionComponent<DeclarativeChartProps> = 
   DeclarativeChartProps
 >((props, forwardedRef) => {
   const { plotlySchema } = sanitizeJson(props.chartSchema);
-  const plotlyInput = plotlySchema as PlotlySchema;
+  const plotlyInput: PlotlySchema | null = (() => {
+    try {
+      return plotlySchema as PlotlySchema;
+    } catch (error) {
+      throw new Error(`Invalid plotly schema: ${error}`);
+    }
+  })();
+
+  if (!plotlyInput) {
+    throw new Error('Plotly input is null or undefined');
+  }
+  if (!plotlyInput.data) {
+    throw new Error('Plotly input data is null or undefined');
+  }
+  if (plotlyInput.data.length === 0) {
+    throw new Error('Plotly input data is empty');
+  }
   let { selectedLegends } = plotlySchema;
   const colorMap = useColorMapping();
   const theme = useTheme();
@@ -223,7 +239,10 @@ export const DeclarativeChart: React.FunctionComponent<DeclarativeChartProps> = 
       );
     case 'bar':
       const orientation = plotlyInput.data[0].orientation;
-      if (orientation === 'h' && isNumberArray((plotlyInput.data[0] as PlotData).x)) {
+      const containsBase = plotlyInput.data.some((series: PlotData) => series.base !== undefined);
+      if (orientation === 'h' && containsBase) {
+        throw new Error('Unsupported chart type: Gantt');
+      } else if (orientation === 'h' && isNumberArray((plotlyInput.data[0] as PlotData).x)) {
         return (
           <ResponsiveHorizontalBarChartWithAxis
             {...transformPlotlyJsonToHorizontalBarWithAxisProps(plotlySchema, colorMap, isDarkTheme)}
@@ -300,6 +319,8 @@ export const DeclarativeChart: React.FunctionComponent<DeclarativeChartProps> = 
           {...commonProps}
         />
       );
+    case 'contour':
+      throw new Error(`Unsupported chart - type: ${plotlyInput.data[0]?.type}`);
     default:
       const xValues = (plotlyInput.data[0] as PlotData).x;
       const yValues = (plotlyInput.data[0] as PlotData).y;