+++ packages/charts/react-charting/CHANGELOG.md
@@ -1,8 +1,30 @@
 # Change Log - @fluentui/react-charting
 
-This log was last generated on Thu, 13 Mar 2025 07:20:52 GMT and should not be manually modified.
+This log was last generated on Tue, 18 Mar 2025 07:21:01 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [5.23.61](https://github.com/microsoft/fluentui/tree/@fluentui/react-charting_v5.23.61)
+
+Tue, 18 Mar 2025 07:21:01 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-charting_v5.23.60..@fluentui/react-charting_v5.23.61)
+
+### Patches
+
+- fix(react-charting): Handle unsupported chart types Gantt and Contour ([PR #34005](https://github.com/microsoft/fluentui/pull/34005) by 120183316+srmukher@users.noreply.github.com)
+
+## [5.23.60](https://github.com/microsoft/fluentui/tree/@fluentui/react-charting_v5.23.60)
+
+Fri, 14 Mar 2025 07:23:15 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-charting_v5.23.59..@fluentui/react-charting_v5.23.60)
+
+### Patches
+
+- feat: export legend data and config from Legends component ([PR #33847](https://github.com/microsoft/fluentui/pull/33847) by 110246001+krkshitij@users.noreply.github.com)
+- fix(react-charting): Handle invalid plotly data ([PR #33982](https://github.com/microsoft/fluentui/pull/33982) by 120183316+srmukher@users.noreply.github.com)
+- Bump @fluentui/react-focus to v8.9.22 ([PR #33981](https://github.com/microsoft/fluentui/pull/33981) by beachball)
+- Bump @fluentui/theme-samples to v8.7.201 ([PR #33981](https://github.com/microsoft/fluentui/pull/33981) by beachball)
+- Bump @fluentui/react to v8.122.12 ([PR #33981](https://github.com/microsoft/fluentui/pull/33981) by beachball)
 
 ## [5.23.59](https://github.com/microsoft/fluentui/tree/@fluentui/react-charting_v5.23.59)